mesa-lib/libva update.
Updating mesa-lib to 9.0, libva to 1.0.16 (bugfix).

After the libdrm update, this works on several of my systems including
NV cards. I will test libva-1.0 and an Intel gfx system next.

Extended message describing optimal choices for libva/mesa/drm. This
message will play every time this compiles so people are reminded
of what Gallium actually is in comparison with DRI.

Note: Gallium performance in GenX is poor and bug-ridden, you do not
want Gallium on Intel onboard GFX, ever. It's just not worth it,
and Intel does not support the gallium Intel drivers.