mesa-lib: module refactoring to solve the impossible option combinations
This is one module that is just too complex for our current system to
make fully optional driven. I counted somewhere between 10-15 combinations
that resulted in build failure. After the refactoring it will always
compile successfully. It does add a little extra compile time due to a
couple of extra dependenices, but that little overhead over frustration
is an easy win.